+++ packages/expo-modules-autolinking/src/index.ts
@@ -1,5 +1,6 @@
 import commander from 'commander';
 
+import { patchReactImportsAsync } from './ReactImportsPatcher';
 import {
   findModulesAsync,
   resolveModulesAsync,
@@ -54,6 +55,15 @@ function registerResolveCommand<OptionsType extends ResolveOptions>(
   return registerSearchCommand<OptionsType>(commandName, fn);
 }
 
+// Register for `patch-react-imports` command
+function registerPatchReactImportsCommand() {
+  return commander
+    .command('patch-react-imports [paths...]')
+    .requiredOption('--pods-root <podsRoot>', 'The path to `Pods` directory')
+    .option('--dry-run', 'Only list files without writing changes to the file system')
+    .action(patchReactImportsAsync);
+}
+
 module.exports = async function (args: string[]) {
   // Searches for available expo modules.
   registerSearchCommand<SearchOptions & { json?: boolean }>('search', async (results, options) => {
@@ -101,6 +111,8 @@ module.exports = async function (args: string[]) {
       'Whether to only generate an empty list. Might be used when the user opts-out of autolinking.',
       false
     );
+
+  registerPatchReactImportsCommand();
 
   await commander
     .version(require('expo-modules-autolinking/package.json').version)