+++ packages/expo-modules-autolinking/src/types.ts
@@ -24,6 +24,11 @@ export interface GenerateOptions extends ResolveOptions {
   empty?: boolean;
 }
 
+export interface PatchReactImportsOptions {
+  podsRoot: string;
+  dryRun: boolean;
+}
+
 export type PackageRevision = {
   path: string;
   version: string;